banner_app/requirements.txt
absl-py==0.7.1
albumentations==0.4.2
alembic==1.0.11
amqp==2.5.1
aniso8601==7.0.0
astor==0.8.0
backcall==0.1.0
beautifulsoup4==4.8.1
billiard==3.6.1.0
blinker==1.4
blis==0.4.1
Bottleneck==1.2.1
cairocffi==1.0.2
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cycler==0.10.0
cymem==2.0.2
dataclasses==0.7
decorator==4.4.0
Django==2.2.6
django-celery-results==1.1.2
django-crispy-forms==1.8.0
django-debug-permissions==1.0.0
dominate==2.3.5
editdistance==0.5.3
fastai==1.0.59
fastprogress==0.1.21
gast==0.2.2
google-pasta==0.1.7
graphviz==0.11.1
grpcio==1.22.0
gunicorn==19.9.0
h5py==2.9.0
idna==2.8
imageai==2.1.5
imageio==2.5.0
imgaug==0.2.6
importlib-metadata==0.23
ipython==7.6.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.14.0
Jinja2==2.10.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
kombu==4.6.5
Mako==1.0.14
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
more-itertools==7.2.0
murmurhash==1.0.2
networkx==2.3
numexpr==2.7.0
numpy==1.16.4
nvidia-ml-py3==7.352.0
opencv-python==4.1.1.26
packaging==19.2
pandas==0.25.2
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
plac==1.1.3
preshed==3.0.2
prompt-toolkit==2.0.9
protobuf==3.8.0
psycopg2==2.8.4
ptyprocess==0.6.0
pycparser==2.19
pydot==1.4.1
pydot-ng==2.0.0
Pygments==2.4.2
pyparsing==2.4.0
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.3
PyWavelets==1.0.3
PyYAML==5.1.1
redis==3.3.11
requests==2.22.0
scikit-image==0.15.0
scipy==1.3.0
six==1.12.0
sk-video==1.1.10
soupsieve==1.9.4
spacy==2.2.2
SQLAlchemy==1.3.6
sqlparse==0.3.0
srsly==0.2.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
thinc==7.3.1
torch==1.3.0
torchvision==0.4.1
tqdm==4.37.0
traitlets==4.3.2
urllib3==1.25.3
vine==1.3.0
visitor==0.1.3
wasabi==0.3.0
wcwidth==0.1.7
Werkzeug==0.15.4
wrapt==1.11.2
WTForms==2.2.1
zipp==0.6.0
